buildtool: support for building programs specified with asterisks
Available only if scons is one of the targets and only if all target
directories have same architecture (list of targets is established for
one of them so it would cause problems if allowed many).

Implemented using asking scons variant build for a list of targets it
finds - it is needed to know enabled repositories and specs to be able
to specify list of valid targets.