genode_target: centralized dependencies handling
Active targets are those which:
 * are required directly or transitively as a dependency
 * is not disabled

Target disabling is automatically propagated to targets that depend on
disabled target.

Target usage counts are automatically handled for dependent targets.

Special GenodeAll target is introduced to be a container for targets to
be built. It is special as its usage count is one initially and cannot
be disabled even if some dependent targets are disabled.