mkevaluator: implemented skipping handling rules
Up until recently rules were not haneled at all and for some targets
there are overlays that implement custom targets from python code.

Now when rules are interpreted there is an issue with multiply more
than one version of action that generates the same target. For this
reason temporary mechanism to skip rule processing is implemented
until overlays are removed.